requirements.txt
dj-database-url==0.5.0
Django==2.2.4
gunicorn==19.9.0
psycopg2==2.8.3
python-dotenv==0.10.3
whitenoise==4.1.3
